+++ cli/connect/connect.go
@@ -33,6 +33,8 @@ type ConnectCtx struct {
 	SslCiphers string
 	// Interactive mode is used.
 	Interactive bool
+	// ConnectTarget contains connection target string: URI or instance name.
+	ConnectTarget string
 }
 
 const (
@@ -70,7 +72,7 @@ func getEvalCmd(connectCtx ConnectCtx) (string, error) {
 
 // Connect establishes a connection to the instance and starts the console.
 func Connect(connectCtx ConnectCtx, connOpts connector.ConnectOpts) error {
-	if err := runConsole(connOpts, "", connectCtx.Language); err != nil {
+	if err := runConsole(connOpts, connectCtx, ""); err != nil {
 		return fmt.Errorf("failed to run interactive console: %s", err)
 	}
 	return nil
@@ -124,8 +126,8 @@ func Eval(connectCtx ConnectCtx, connOpts connector.ConnectOpts, args []string) 
 }
 
 // runConsole run a new console.
-func runConsole(connOpts connector.ConnectOpts, title string, lang Language) error {
-	console, err := NewConsole(connOpts, title, lang)
+func runConsole(connOpts connector.ConnectOpts, connectCtx ConnectCtx, title string) error {
+	console, err := NewConsole(connOpts, connectCtx, title)
 	if err != nil {
 		return fmt.Errorf("failed to create new console: %s", err)
 	}